requirements.txt
Django>=4.2
djangorestframework>=3.14.0
psycopg2>=2.9.6
celery>=5.3.0
redis>=4.6.0
requests>=2.28.0
python-decouple>=3.8
gunicorn>=20.1.0
django-cors-headers>=3.14.0
python-dotenv>=1.0
django-celery-beat>=2.7.0
